proposer update
