+++ client/attester/BUILD.bazel
@@ -8,11 +8,7 @@ go_library(
     visibility = ["//client:__subpackages__"],
     deps = [
         "//client/types:go_default_library",
-        "//proto/beacon/rpc/v1:go_default_library",
-        "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
-        "@io_bazel_rules_go//proto/wkt:empty_go_proto",
-        "@org_golang_x_crypto//blake2b:go_default_library",
     ],
 )
 
@@ -24,14 +20,10 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//client/internal:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",  #keep
-        "//proto/beacon/rpc/v1:go_default_library",
         "//shared/testutil:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",  #keep
-        "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
-        "@io_bazel_rules_go//proto/wkt:empty_go_proto",
         "@org_golang_google_grpc//:go_default_library",  #keep
         "@org_golang_google_grpc//metadata:go_default_library",  #keep
     ],
