+++ client/beacon-client/BUILD.bazel
@@ -3,22 +3,30 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["service.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client/proposer",
+    importpath = "github.com/prysmaticlabs/prysm/client/beacon-client",
     visibility = ["//client:__subpackages__"],
     deps = [
         "//client/types:go_default_library",
+        "//proto/beacon/rpc/v1:go_default_library",
+        "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
+        "@io_bazel_rules_go//proto/wkt:empty_go_proto",
+        "@org_golang_x_crypto//blake2b:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    size = "small",
     srcs = ["service_test.go"],
     embed = [":go_default_library"],
     deps = [
+        "//client/internal:go_default_library",
+        "//proto/beacon/p2p/v1:go_default_library",
+        "//proto/beacon/rpc/v1:go_default_library",
         "//shared/testutil:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
+        "@io_bazel_rules_go//proto/wkt:empty_go_proto",
     ],
 )
