+++ client/node/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
     visibility = ["//client:__subpackages__"],
     deps = [
         "//client/attester:go_default_library",
+        "//client/beacon-client:go_default_library",
         "//client/mainchain:go_default_library",
         "//client/params:go_default_library",
         "//client/proposer:go_default_library",
